// src/webapp/halshow/src/generated/halcmd_client.ts
export interface CmdResult {
  success: boolean;
  output?: string | undefined;
  error?: string | undefined;
}

export interface PinInfo {
  name: string;
  type: string;
  dir: string;
  value: string;
  owner: string;
  linked: boolean;
  signal?: string | undefined;
  alias?: string | undefined;
}

export interface ParamInfo {
  name: string;
  type: string;
  dir: string;
  value: string;
  owner: string;
  alias?: string | undefined;
}

export interface SignalInfo {
  name: string;
  type: string;
  value: string;
  writers: string[];
  readers: string[];
  bidirs: string[];
}

export interface ComponentInfo {
  name: string;
  id: number;
  type: string;
  state: string;
  pid?: number | undefined;
  timer?: number | undefined;
}

export interface FunctionInfo {
  name: string;
  owner: string;
  users: number;
  runtime: number;
  fp: boolean;
}

export interface ThreadInfo {
  name: string;
  period: number;
  fp: boolean;
  functions: string[];
}

export interface HalStatus {
  rt_lock: boolean;
  mem_lock: boolean;
  threads_running: boolean;
  components: number;
  pins: number;
  signals: number;
  params: number;
  threads: number;
  functions: number;
}